requirements.txt
Flask
flask-cors
python-docx
PyMuPDF
gunicorn
openai